Resolve two minor Windows specific corectness warnings found by a newer compiler version.
* subversion/libsvn_subr/io.c
  (svn_io__utf8_to_unicode_longpath): Remove unnecessary truncate.

* subversion/libsvn_subr/win32_crashrpt.c
  (write_value): Properly print pointersized value.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1764436 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_subr/io.c
@@ -1774,7 +1774,7 @@ svn_io__utf8_to_unicode_longpath(const WCHAR **result,
      * than the original number of utf-8 narrow chars.
      */
     const WCHAR *prefix = NULL;
-    const int srclen = strlen(source);
+    const size_t srclen = strlen(source);
     WCHAR *buffer;
 
     if (srclen > 248)
